+++ HAL/modules/defaults/display/focusOnFit2D.py
@@ -10,20 +10,17 @@ Comments : a basic (2D) image display. Can be used as an example when building
 # %% IMPORTS
 
 # -- global
-import logging
 import pyqtgraph as pg
 import numpy as np
 
 # -- local
-from .abstractImage import AbstractImageDisplay
+from HAL.classes.display.abstractImage import AbstractImageDisplay
 
-# -- logger
-logger = logging.getLogger(__name__)
 
 # %% CLASS DEFINITION
 
 
-class BasicImageDisplay(AbstractImageDisplay):
+class FocusOnFit2D(AbstractImageDisplay):
     """a basic (2D) image display
     NB : inherits a LOT of methods from AbstractImageDisplay().
          Have a look at this class if you do not find a method here !
@@ -46,7 +43,7 @@ class BasicImageDisplay(AbstractImageDisplay):
         self._current_levels = (0, 1)
         self._data_in = None
         # -- other attributes
-        self.name = "Basic 2D"
+        self.name = "Focus on fit 2D"
 
     # -- DISPLAY MANAGEMENT
 
@@ -56,16 +53,14 @@ class BasicImageDisplay(AbstractImageDisplay):
         self.screen.clear()
 
         # -- create plot layout
-        # add plots
-        im_plot = self.screen.addPlot(1, 1)
-        fit_plot = self.screen.addPlot(2, 2)
-        cx_plot = self.screen.addPlot(2, 1)
-        cy_plot = self.screen.addPlot(1, 2)
+        roi_plot = self.screen.addPlot(1, 1, title="Data (ROI)")
+        fit_plot = self.screen.addPlot(2, 1, title="Fit (ROI)")
+        err_plot = self.screen.addPlot(2, 2, title="Error")
+        im_plot = self.screen.addPlot(1, 2, title="Data (all)")
 
         # stretch
         layout = self.screen.ci.layout
-        layout.setColumnStretchFactor(1, 3)
-        layout.setRowStretchFactor(1, 3)
+        layout.setColumnStretchFactor(1, 1)
 
         # configure subplot
         im_plot.setAspectLocked(lock=True, ratio=1)
@@ -75,8 +70,8 @@ class BasicImageDisplay(AbstractImageDisplay):
         # store
         self.image_plot = im_plot
         self.fit_plot = fit_plot
-        self.cx_plot = cx_plot
-        self.cy_plot = cy_plot
+        self.roi_plot = roi_plot
+        self.err_plot = err_plot
 
         # -- init image item
         # main image
@@ -84,10 +79,20 @@ class BasicImageDisplay(AbstractImageDisplay):
         im_plot.addItem(img)
         self.current_image = img
 
+        # roi image
+        roi_img = pg.ImageItem()
+        roi_plot.addItem(roi_img)
+        self.current_roi_image = roi_img
+
         # fit image
         fit_img = pg.ImageItem()
         fit_plot.addItem(fit_img)
         self.current_fit_image = fit_img
+
+        # err image
+        err_img = pg.ImageItem()
+        err_plot.addItem(err_img)
+        self.current_err_image = err_img
 
     def clearPlot(self):
         """clear all the data plot"""
@@ -129,6 +134,17 @@ class BasicImageDisplay(AbstractImageDisplay):
         # set colormap
         self.updateColormap(colormap)
 
+        # update roi (so that the roi is refreshed even if no fit is loaded)
+        if selected_ROI is not None:
+            Z, _ = self.getROIData(selected_ROI)
+            if Z is not None:
+                # roi
+                self.current_roi_image.updateImage(image=Z, levels=levels)
+                self.updateColormap(
+                    colormap=colormap,
+                    image=self.current_roi_image,
+                )
+
     def BackgroundChangedFinished(self):
         """triggered when the background area was moved"""
         if self._data_in is None:
@@ -149,11 +165,10 @@ class BasicImageDisplay(AbstractImageDisplay):
         )
 
     def clearFit(self):
-        """resets fit display"""
-        self.cx_plot.clear()
-        self.cy_plot.clear()
         image = np.zeros((10, 10))
         self.current_fit_image.updateImage(image=image, levels=(0, 1))
+        self.current_err_image.updateImage(image=image, levels=(0, 1))
+        self.err_plot.setTitle("Error")
 
     def updateFit(
         self,
@@ -172,27 +187,27 @@ class BasicImageDisplay(AbstractImageDisplay):
         fit = fit_dic[selected_ROI]
         Zfit = fit.eval((X, Y))
 
-        # -- plot cuts
-        # coordinates cuts
-        x = X[:, 0]
-        y = Y[0, :]
-
-        # get argmax
-        i, j = np.unravel_index(Zfit.argmax(), Zfit.shape)
-
-        # plot x cut
-        self.cx_plot.clear()
-        self.cx_plot.plot(x, Z[:, j], symbol="+", size=1, symbolSize=5)
-        self.cx_plot.plot(x, Zfit[:, j], pen=pg.mkPen(color=(255, 0, 0), width=2))
-
-        # plot y cut
-        self.cy_plot.clear()
-        self.cy_plot.plot(Z[i, :], y, symbol="+", size=1, symbolSize=5)
-        self.cy_plot.plot(Zfit[i, :], y, pen=pg.mkPen(color=(255, 0, 0), width=2))
+        # error
+        Zerr = Z - Zfit
 
         # -- display
+        # roi
+        self.current_roi_image.updateImage(image=Z, levels=self._current_levels)
+        self.updateColormap(
+            colormap=self._current_colormap, image=self.current_roi_image
+        )
+
         # fit
         self.current_fit_image.updateImage(image=Zfit, levels=self._current_levels)
         self.updateColormap(
             colormap=self._current_colormap, image=self.current_fit_image
+        )
+
+        # err
+        err_ampl = np.max(np.abs(Zerr))
+        self.current_err_image.updateImage(image=Zerr, levels=(-err_ampl, err_ampl))
+        new_title = "Error : range = ± %.2g" % err_ampl
+        self.err_plot.setTitle(new_title)
+        self.updateColormap(
+            colormap="RdBu", image=self.current_err_image, update_current=False
         )
